ass4 soort van af?????

+++ Assignment3/Assignment3/kernel3.cl
@@ -2,7 +2,7 @@ kernel void kernel3(global int *gdata, local int *sdata) {
 	unsigned int tid = get_local_id(0);
 	unsigned int gid = get_global_id(0);
 	unsigned int lsz = get_local_size(0);
-	unsigned long long int i = get_group_id(0) * lsz * 2 + tid;
+	unsigned long int i = get_group_id(0) * lsz * 2 + tid;
 
 	// copy int’s from Global to Local memory
 	sdata[tid] = gdata[i] + gdata[i + lsz];
